Allacrity: fix terminal overflowing w/ Vim
Setting an arbitrarily high value for lines and columns forces Allacrity
to start maximized and the same could be accomplished by using the
'maximized' flag. Problem is that with the first Neovim overflow and with
the latter underflow requiring a maximized state toogle every time I
open it so to fix the flow. My solution was to choose a sensible value
for lines and columns which almost maximizes the window without
overflowing it.